.history/requirements_20250214235741.txt
streamlit>=1.0.0
Pillow
langchain
sentence-transformers
faiss-cpu
python-dotenv
plotly
pandas
transformers
